+++ analytics/utils/department_analytic_helpers.py
@@ -1,3 +1,7 @@
+from analytics.utils.system_overview_helpers.assets import build_department_accessory_trends, build_department_asset_trends
+from analytics.utils.system_overview_helpers.kpis import build_department_kpis
+from analytics.utils.system_overview_helpers.returns import build_department_return_flow_trends, build_department_return_state_trends
+from analytics.utils.system_overview_helpers.valuation import build_department_asset_value_trends
 from sites.models.sites import Department
 from datetime import timedelta
 
@@ -8,124 +12,6 @@ from analytics.models.snapshots import DailyDepartmentSnapshot
 from analytics.utils.analytics_helpers import percentage_delta, truncate_date
 from analytics.utils.utils.viewset_helpers import get_snapshot_range_start
 
-def build_department_kpis(*, department: Department):
-    today = (
-        DailyDepartmentSnapshot.objects
-        .filter(department=department)
-        .order_by("-snapshot_date")
-        .first()
-    )
-
-    yesterday = (
-        DailyDepartmentSnapshot.objects
-        .filter(
-            department=department,
-            snapshot_date=today.snapshot_date - timedelta(days=1),
-        )
-        .first()
-        if today else None
-    )
-
-    def safe_delta(current, previous):
-        return percentage_delta(current, previous) if current is not None else None
-
-    if not today:
-        return {}
-
-    return {
-        # People
-        "total_users": {
-            "value": today.total_users,
-            "delta": safe_delta(
-                today.total_users,
-                yesterday.total_users if yesterday else None,
-            ),
-        },
-        "total_admins": {
-            "value": today.total_admins,
-            "delta": safe_delta(
-                today.total_admins,
-                yesterday.total_admins if yesterday else None,
-            ),
-        },
-
-        # Equipment
-        "total_equipment": {
-            "value": today.total_equipment,
-            "delta": safe_delta(
-                today.total_equipment,
-                yesterday.total_equipment if yesterday else None,
-            ),
-        },
-        "equipment_ok": {
-            "value": today.equipment_ok,
-            "delta": safe_delta(
-                today.equipment_ok,
-                yesterday.equipment_ok if yesterday else None,
-            ),
-        },
-        "equipment_under_repair": {
-            "value": today.equipment_under_repair,
-            "delta": safe_delta(
-                today.equipment_under_repair,
-                yesterday.equipment_under_repair if yesterday else None,
-            ),
-        },
-        "equipment_damaged": {
-            "value": today.equipment_damaged,
-            "delta": safe_delta(
-                today.equipment_damaged,
-                yesterday.equipment_damaged if yesterday else None,
-            ),
-        },
-
-        # Consumables
-        "total_consumables": {
-            "value": today.total_consumables,
-            "delta": safe_delta(
-                today.total_consumables,
-                yesterday.total_consumables if yesterday else None,
-            ),
-        },
-        "total_consumables_quantity": {
-            "value": today.total_consumables_quantity,
-            "delta": safe_delta(
-                today.total_consumables_quantity,
-                yesterday.total_consumables_quantity if yesterday else None,
-            ),
-        },
-        # -------------------------
-        # Returns
-        # -------------------------
-        "total_return_requests": {
-            "value": today.total_return_requests,
-            "delta": safe_delta(
-                today.total_return_requests,
-                yesterday.total_return_requests if yesterday else None,
-            ),
-        },
-        "pending_return_requests": {
-            "value": today.pending_return_requests,
-            "delta": safe_delta(
-                today.pending_return_requests,
-                yesterday.pending_return_requests if yesterday else None,
-            ),
-        },
-        "returns_created_24h": {
-            "value": today.returns_created_last_24h,
-            "delta": safe_delta(
-                today.returns_created_last_24h,
-                yesterday.returns_created_last_24h if yesterday else None,
-            ),
-        },
-        "returns_processed_24h": {
-            "value": today.returns_processed_last_24h,
-            "delta": safe_delta(
-                today.returns_processed_last_24h,
-                yesterday.returns_processed_last_24h if yesterday else None,
-            ),
-        },
-    }
 
 def build_department_consumable_trends(*, department, days, granularity):
     start = get_snapshot_range_start(
@@ -172,51 +58,6 @@ def build_department_consumable_trends(*, department, days, granularity):
         for row in qs
     ]
 
-def build_department_accessory_trends(*, department, days, granularity):
-    start = get_snapshot_range_start(
-        model=DailyDepartmentSnapshot,
-        days=days,
-        date_field="snapshot_date",
-        filters={"department": department},
-    )
-
-    if not start:
-        return []
-
-    base = (
-        DailyDepartmentSnapshot.objects
-        .filter(
-            department=department,
-            snapshot_date__gte=start,
-        )
-        .annotate(period=truncate_date("snapshot_date", granularity))
-    )
-
-    latest_per_period = (
-        base.values("period")
-        .annotate(latest_date=Max("snapshot_date"))
-    )
-
-    qs = (
-        base.filter(
-            snapshot_date=Subquery(
-                latest_per_period
-                .filter(period=OuterRef("period"))
-                .values("latest_date")[:1]
-            )
-        )
-        .order_by("period")
-    )
-
-    return [
-        {
-            "date": row.period.isoformat(),
-            "total_accessories": row.total_accessories,
-            "total_accessories_quantity": row.total_accessories_quantity,
-        }
-        for row in qs
-    ]
-
 def build_department_user_trends(*, department, days, granularity):
     start = get_snapshot_range_start(
         model=DailyDepartmentSnapshot,
@@ -259,199 +100,9 @@ def build_department_user_trends(*, department, days, granularity):
         for row in qs
     ]
 
-def build_department_asset_trends(*, department, days, granularity):
-    start = get_snapshot_range_start(
-    model=DailyDepartmentSnapshot,
-    days=days,
-    date_field="snapshot_date",
-    filters={"department": department},
-)
 
-    if not start:
-        return []
 
-    base = (
-        DailyDepartmentSnapshot.objects
-        .filter(department=department, snapshot_date__gte=start)
-        .annotate(period=truncate_date("snapshot_date", granularity))
-    )
 
-    latest_per_period = (
-        base.values("period")
-        .annotate(latest_date=Max("snapshot_date"))
-    )
-
-    qs = (
-        base.filter(
-            snapshot_date=Subquery(
-                latest_per_period
-                .filter(period=OuterRef("period"))
-                .values("latest_date")[:1]
-            )
-        )
-        .order_by("period")
-    )
-
-    return [
-        {
-            "date": row.period.isoformat(),
-            "equipment_ok": row.equipment_ok,
-            "equipment_under_repair": row.equipment_under_repair,
-            "equipment_damaged": row.equipment_damaged,
-        }
-        for row in qs
-    ]
-def build_department_accessory_trends(*, department, days, granularity):
-    start = get_snapshot_range_start(
-        model=DailyDepartmentSnapshot,
-        days=days,
-        date_field="snapshot_date",
-        filters={"department": department},
-    )
-
-    if not start:
-        return []
-
-    base = (
-        DailyDepartmentSnapshot.objects
-        .filter(
-            department=department,
-            snapshot_date__gte=start,
-        )
-        .annotate(period=truncate_date("snapshot_date", granularity))
-    )
-
-    latest_per_period = (
-        base.values("period")
-        .annotate(latest_date=Max("snapshot_date"))
-    )
-
-    qs = (
-        base.filter(
-            snapshot_date=Subquery(
-                latest_per_period
-                .filter(period=OuterRef("period"))
-                .values("latest_date")[:1]
-            )
-        )
-        .order_by("period")
-    )
-
-    return [
-        {
-            "date": row.period.isoformat(),
-            "total_accessories": row.total_accessories,
-            "total_accessories_quantity": row.total_accessories_quantity,
-        }
-        for row in qs
-    ]
-
-def build_department_return_state_trends(*, department, days, granularity):
-    start = get_snapshot_range_start(
-        model=DailyDepartmentSnapshot,
-        days=days,
-        date_field="snapshot_date",
-        filters={"department": department},
-    )
-
-    if not start:
-        return []
-
-    base = (
-        DailyDepartmentSnapshot.objects
-        .filter(department=department, snapshot_date__gte=start)
-        .annotate(period=truncate_date("snapshot_date", granularity))
-    )
-
-    latest_per_period = (
-        base.values("period")
-        .annotate(latest_date=Max("snapshot_date"))
-    )
-
-    qs = (
-        base.filter(
-            snapshot_date=Subquery(
-                latest_per_period
-                .filter(period=OuterRef("period"))
-                .values("latest_date")[:1]
-            )
-        )
-        .order_by("period")
-    )
-
-    return [
-        {
-            "date": row.period.isoformat(),
-            "pending": row.pending_return_requests,
-            "approved": row.approved_return_requests,
-            "denied": row.denied_return_requests,
-            "partial": row.partial_return_requests,
-        }
-        for row in qs
-    ]
-
-def build_department_return_flow_trends(*, department, days, granularity):
-    start = get_snapshot_range_start(
-        model=DailyDepartmentSnapshot,
-        days=days,
-        date_field="snapshot_date",
-        filters={"department": department},
-    )
-
-    if not start:
-        return []
-
-    qs = (
-        DailyDepartmentSnapshot.objects
-        .filter(department=department, snapshot_date__gte=start)
-        .annotate(period=truncate_date("snapshot_date", granularity))
-        .values("period")
-        .annotate(
-            created=Sum("returns_created_last_24h"),
-            processed=Sum("returns_processed_last_24h"),
-        )
-        .order_by("period")
-    )
-
-    return [
-        {
-            "date": row["period"].isoformat(),
-            "created": row["created"],
-            "processed": row["processed"],
-        }
-        for row in qs
-    ]
-def build_department_return_flow_trends(*, department, days, granularity):
-    start = get_snapshot_range_start(
-        model=DailyDepartmentSnapshot,
-        days=days,
-        date_field="snapshot_date",
-        filters={"department": department},
-    )
-
-    if not start:
-        return []
-
-    qs = (
-        DailyDepartmentSnapshot.objects
-        .filter(department=department, snapshot_date__gte=start)
-        .annotate(period=truncate_date("snapshot_date", granularity))
-        .values("period")
-        .annotate(
-            created=Sum("returns_created_last_24h"),
-            processed=Sum("returns_processed_last_24h"),
-        )
-        .order_by("period")
-    )
-
-    return [
-        {
-            "date": row["period"].isoformat(),
-            "created": row["created"],
-            "processed": row["processed"],
-        }
-        for row in qs
-    ]
 
 
 def get_department_overview(*, department, days, granularity, sections):
@@ -497,6 +148,15 @@ def get_department_overview(*, department, days, granularity, sections):
             department=department,
             days=days,
             granularity=granularity,
+        )
+    
+    if "asset_value" in sections:
+        charts["asset_value"] = (
+            build_department_asset_value_trends(
+                department=department,
+                days=days,
+                granularity=granularity,
+            )
         )
